# feat(auth): implement core authentication boilerplate - Task 16