--- a/apps/auth-service/src/app/app.controller.ts
+++ b/apps/auth-service/src/app/app.controller.ts
@@ -1,6 +1,7 @@
 import { Controller } from '@nestjs/common';
 import { EXCHANGES, AUTH_ROUTING_KEYS } from '@usecapsule/messaging';
 import { RabbitRPC, RabbitPayload } from '@usecapsule/rabbitmq';
+import { AuthRateLimit } from '@usecapsule/shared-redis';
 import type { HealthCheckResponse } from '@usecapsule/types';
 
 import type { RegisterUserCommand } from '../modules/user-management/commands/register-user.command';
@@ -51,6 +52,17 @@ export class AppController {
    * @param dto - User registration data from the request
    * @returns Promise resolving to registration response with user data
    */
+  @AuthRateLimit({
+    limit: 5,
+    windowSize: 300, // 5 minutes
+    blockDuration: 900, // 15 minutes
+    keyGenerator: (context: { body?: { email?: string }; ip?: string }) => {
+      const email = context.body?.email as string;
+      const ip = context.ip || 'unknown';
+      return email ? `auth:register:${email}` : `auth:register:${ip}`;
+    },
+    message: 'Too many registration attempts. Please try again later.',
+  })
   @RabbitRPC({
     exchange: EXCHANGES.COMMANDS,
     routingKey: AUTH_ROUTING_KEYS.REGISTER,