--- a/libs/shared/ddd/src/index.ts
+++ b/libs/shared/ddd/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/ddd.module';
+export * from './lib/entities';