… .claude/agents/ddd-architect.md
… .github/workflows/ci.yml
… .github/workflows/claude-code-review.yml
… .taskmaster/state.json
… .taskmaster/tasks/task_016_feature-auth-core-implementation.txt
… .taskmaster/tasks/tasks.json
… CLAUDE.md
… TASKS.md
… apps/auth-service/database/migrations/V1.4.0__Create_email_verification_tokens_table.sql
… apps/auth-service/database/migrations/V1.5.0__Create_password_reset_tokens_table.sql
… apps/auth-service/database/migrations/V1.6.0__Create_two_factor_auth_table.sql
… apps/auth-service/database/migrations/V1.7.0__Create_oauth_providers_table.sql
… apps/auth-service/database/migrations/V1.8.0__Create_api_keys_table.sql
… apps/auth-service/database/migrations/V1.9.0__Create_auth_audit_logs_table.sql
… apps/auth-service/src/app/app.module.ts
… apps/auth-service/src/modules/auth/auth.module.ts
… apps/auth-service/src/modules/auth/dto/jwt-payload.interface.ts
… apps/auth-service/src/modules/auth/dto/token.dto.ts
… apps/auth-service/src/modules/auth/services/jwt.service.ts
… apps/auth-service/src/modules/auth/services/password.service.ts
… apps/auth-service/src/modules/auth/services/token-blacklist.service.ts
… apps/auth-service/src/modules/roles/domain/role.entity.ts
… apps/auth-service/src/modules/roles/guards/permissions.decorator.ts
… apps/auth-service/src/modules/roles/guards/roles.guard.ts
… apps/auth-service/src/modules/roles/roles.module.ts
… apps/auth-service/src/modules/roles/services/role.service.ts
… apps/auth-service/src/modules/user-management/commands/register-user.command.ts
… apps/auth-service/src/modules/user-management/database/user.repository.interface.ts
… apps/auth-service/src/modules/user-management/database/user.repository.ts
… apps/auth-service/src/modules/user-management/domain/entities/user.entity.ts