# 🔐 Core Authentication Boilerplate Implementation

This PR implements the foundational authentication system for the Capsule Platform, corresponding to **Task 16** in TaskMaster.

## 📋 Overview

Implements a complete authentication system using Domain-Driven Design principles within the existing microservices architecture. All authentication operations communicate via RabbitMQ using exchange-based routing with the API Gateway.

## 🎯 TaskMaster Reference

- **Task**: 16 - Implement Core Authentication Boilerplate  
- **Tag**: `feature-auth-core-implementation`
- **Priority**: HIGH
- **Dependencies**: None (foundational task)

## 🔗 Related GitHub Issues

- [x] #6 - Auth Service Base Implementation *(infrastructure complete)*
- [ ] #7 - JWT Token Generation and Validation *(subtask 16.2)*
- [ ] #16 - User Registration and Login *(subtasks 16.3, 16.4)*  
- [ ] #17 - Password Reset Flow *(subtask 16.5)*
- [ ] #18 - RBAC Implementation *(subtask 16.6)*
- [ ] #37 - Authentication Middleware *(subtask 16.10)*

## 🏗️ Implementation Plan

### ✅ Already Completed
- **16.1**: Database schemas and entity models - Complete Zod schemas in `libs/configs/database/src/lib/schemas/auth.schemas.ts`
- **16.9**: RabbitMQ message handlers - @golevelup/nestjs-rabbitmq integration with routing keys

### 🚧 To Be Implemented

#### 🔑 **16.2 - Password Hashing & JWT Services**
```typescript
// New files to create:
apps/auth-service/src/modules/auth/services/
├── password.service.ts        # bcrypt hashing with salt rounds
├── jwt.service.ts            # JWT generation/validation  
└── token.service.ts          # Token management & blacklist
```

#### 📝 **16.3 - User Registration Handler**  
```typescript  
// New files to create:
apps/auth-service/src/modules/user/commands/
├── register-user.command.ts
├── register-user.handler.ts
└── register-user.dto.ts
```

#### 🔐 **16.4 - User Login Handler**
```typescript
// New files to create:  
apps/auth-service/src/modules/auth/commands/
├── login-user.command.ts
├── login-user.handler.ts  
└── login-user.dto.ts
```

#### 🔄 **16.5 - Password Reset Flow**
```typescript
// New files to create:
apps/auth-service/src/modules/password-reset/
├── commands/
│   ├── request-password-reset.command.ts
│   └── confirm-password-reset.command.ts
└── services/
    └── password-reset.service.ts
```

#### 👤 **16.6 - RBAC System**  
```typescript
// New files to create:
apps/auth-service/src/modules/rbac/
├── services/role.service.ts
├── guards/roles.guard.ts
└── decorators/roles.decorator.ts
```

#### ✉️ **16.7 - Email Verification**
```typescript
// New files to create:
apps/auth-service/src/modules/email-verification/
├── commands/
│   ├── verify-email.command.ts
│   └── resend-verification.command.ts  
└── services/
    └── email-verification.service.ts
```

#### 🎫 **16.8 - Session Management**
```typescript
// New files to create:
apps/auth-service/src/modules/session/
├── services/session.service.ts
├── commands/logout.command.ts
└── queries/get-active-sessions.query.ts
```

#### 🛡️ **16.10 - Authentication Middleware**
```typescript  
// New files to create:
apps/api-gateway/src/modules/auth/
├── guards/jwt-auth.guard.ts
├── guards/roles.guard.ts  
├── middleware/auth.middleware.ts
└── decorators/public.decorator.ts
```

## 🗃️ Database Migrations

Will implement Flyway migrations based on existing schemas:

```sql
-- To be created:
apps/auth-service/database/migrations/
├── V1.1.0__Create_users_table.sql
├── V1.2.0__Create_user_roles_table.sql  
├── V1.3.0__Create_auth_sessions_table.sql
├── V1.4.0__Create_verification_tokens_table.sql
├── V1.5.0__Create_password_reset_tokens_table.sql
├── V1.6.0__Create_two_factor_auth_table.sql
├── V1.7.0__Create_oauth_providers_table.sql
├── V1.8.0__Create_api_keys_table.sql
└── V1.9.0__Create_auth_audit_logs_table.sql
```

## 🔄 RabbitMQ Message Patterns

Will implement these RPC handlers using existing routing keys:

```typescript
// Command handlers to implement:
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.register' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.login' })  
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.logout' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.validate-token' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.refresh-token' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.reset-password' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.verify-email' })
@RabbitRPC({ exchange: 'capsule.commands', routingKey: 'auth.check-permissions' })
```

## 🧪 Testing Strategy

### Unit Tests (80%+ coverage target)
- All service methods with mocked dependencies
- Command handlers with isolated business logic
- Password hashing and JWT generation/validation
- RBAC permission checking logic

### Integration Tests  
- RabbitMQ message flow end-to-end
- Database operations with test containers
- Email verification token flows
- Session management across requests

### Security Tests
- JWT token security and expiration
- Password strength validation  
- Rate limiting on auth attempts
- SQL injection prevention
- Cross-site scripting (XSS) prevention

## 📊 Current Implementation Status

| Subtask | Component | Status | Priority |
|---------|-----------|--------|----------|
| 16.1 | Database Schemas | ✅ **DONE** | - |
| 16.2 | JWT & Password Services | ❌ **PENDING** | P0 |  
| 16.3 | User Registration | ❌ **PENDING** | P0 |
| 16.4 | User Login | ❌ **PENDING** | P0 |
| 16.5 | Password Reset | ❌ **PENDING** | P1 |
| 16.6 | RBAC System | ❌ **PENDING** | P1 |
| 16.7 | Email Verification | ❌ **PENDING** | P1 |  
| 16.8 | Session Management | ❌ **PENDING** | P1 |
| 16.9 | RabbitMQ Handlers | ✅ **DONE** | - |
| 16.10 | Auth Middleware | ❌ **PENDING** | P0 |

## 🔧 Technical Architecture

### Microservice Communication Flow
```mermaid
graph LR
    A[API Gateway] -->|HTTP Request| B[Auth Middleware]
    B -->|RabbitMQ RPC| C[Auth Service]  
    C -->|PostgreSQL| D[Auth Database]
    C -->|Redis| E[Session Cache]
    C -->|Events| F[Other Services]
```

### Domain Structure
```
apps/auth-service/src/modules/
├── auth/           # Core auth operations (login, tokens)
├── user/           # User management (registration, profile)  
├── rbac/           # Role-based access control
├── session/        # Session management
├── password-reset/ # Password reset flows
└── email-verification/ # Email verification flows
```

## 🛡️ Security Considerations

- **Password Hashing**: bcrypt with configurable salt rounds (default: 10)
- **JWT Security**: RS256 signing, short-lived access tokens (15m)  
- **Session Management**: Secure token storage with Redis
- **Rate Limiting**: Protection against brute force attacks
- **Audit Logging**: Complete auth action tracking
- **Token Blacklisting**: Immediate token revocation capability

## 📈 Performance Targets

- **Auth Response Time**: < 200ms for token validation
- **Registration Time**: < 500ms end-to-end  
- **Login Time**: < 300ms with valid credentials
- **Database Connections**: Optimized pool settings for auth workload
- **Redis Caching**: Session lookup < 10ms

## 🔄 Deployment Considerations

- **Zero Downtime**: Backward compatible database migrations
- **Environment Variables**: Secure JWT secrets and database credentials
- **Health Checks**: Auth service health monitoring
- **Monitoring**: Authentication metrics and alerting

## 🎯 Definition of Done

- [ ] All 8 pending subtasks implemented and tested
- [ ] Database migrations applied successfully  
- [ ] RabbitMQ integration fully functional
- [ ] Unit tests achieving 80%+ coverage
- [ ] Integration tests passing
- [ ] Security tests validating all auth flows
- [ ] Documentation updated
- [ ] Performance targets met
- [ ] Related GitHub issues can be closed

---

🤖 **Generated by Claude Code** | 📋 **TaskMaster Task 16** | 🏷️ **Tag: feature-auth-core-implementation**